fix(andax/tsunagu): gh_tag pypi crates crates_max crates_newest npm
The above functions are subjected to syntax error when an invalid url is
passed into the functions that breaks the traceback generator since
`_tb()` treats it as they uses engine.eval() which confuses `_tb()` to think
that the eval code is inside the same src file.

A constant `ANDA_VER` is added to the rhai scope too.